Refactoring: split DPLangevin and SimDP into separate hpp/cpp pairs
