--- src/application_dp.cpp
@@ -1,11 +1,0 @@
-// 
-// "Dornic" operator-splitting method of integrating DP-type Langevin equations
-//   adapted from code by Paula Villa Martín and Victor Buendía.
-//
-// CPS 2025-09-02
-// 
-
-#include <pybind11/numpy.h>
-#include <string>
-#include "core.hpp"
-#include "application_dp.hpp"
