+++ src/applications_pybind.cpp
@@ -7,7 +7,8 @@
 // 
 #include <pybind11/numpy.h>
 #include "core.hpp"
-#include "application_dp.hpp"
+#include "application_dplangevin.hpp"
+#include "application_dpsim.hpp"
 // Essential for STL container conversions
 #include <pybind11/stl.h>
