+++ test/test_dp.py
@@ -11,15 +11,16 @@ print()
 sim_dp = dplvn.SimDP(
     linear=1.0, quadratic=2.0, 
     diffusion=0.1, noise=1.0, 
-    t_max=100.0, 
+    t_max=50.0-1e-10, 
     # dx=0.5, dt=0.01
     # grid_dimension=dplvn.D1,
     # grid_size=(4096,),
     grid_dimension=dplvn.D2,
     grid_size=(8,8,),
     # grid_size=(64,64,),
-    grid_topology= dplvn.PERIODIC,
+    grid_topology= dplvn.BOUNDED,
     boundary_condition=dplvn.FLOATING,
+    initial_condition=dplvn.RANDOM_UNIFORM,
     integration_method=dplvn.RUNGE_KUTTA
 )
 print(dir(sim_dp))
